requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
tensorflow
SALib
joblibs